requirements.txt
# Core requirements
flask==2.3.3
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
torch==2.0.1
torchvision==0.15.2

# Backend-specific requirements
# Baseline
sqlite3

# Foundation Model
openai==0.28.0
tiktoken==0.4.0

# Graph
neo4j==5.9.0
networkx==3.1

# Simulation and analysis
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0

# Optional: UI components
# flask-socketio==5.3.4
# flask-wtf==1.1.1
